src/org/howard/edu/lsp/oopfinal/exam/question1/SongsDatabaseTest.java
package org.howard.edu.lsp.oopfinal.exam.question1;

import org.howard.edu.lsp.oopfinal.exam.SongsDatabase;
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;

public class SongsDatabaseTest {
    @Test
    void testAddSong() {
        SongsDatabase db = new SongsDatabase();
        db.addSong("Rap", "Savage");
        assertTrue(db.getSongs("Rap").contains("Savage"));
    }

    @Test
    void testGetGenreOfSong() {
        SongsDatabase db = new SongsDatabase();
        db.addSong("Rap", "Savage");
        assertEquals("Rap", db.getGenreOfSong("Savage"));
    }

    @Test
    void testGetSongs() {
        SongsDatabase db = new SongsDatabase();
        db.addSong("Rap", "Savage");
        db.addSong("Rap", "Gin and Juice");
        assertEquals(2, db.getSongs("Rap").size());
    }
}
